proj/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
APScheduler==3.10.4
asgiref==3.8.1
asyncio==3.4.3
attrs==23.2.0
Automat==22.10.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==42.0.5
cssselect==1.2.0
discord.py==2.3.2
Django==5.0.4
django-allauth==0.63.0
django-apscheduler==0.6.2
django-cors-headers==4.3.1
django-extensions==3.2.3
django-oauth2-provider==0.2.6.1
django-rest-auth==0.9.5
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
filelock==3.13.4
frozenlist==1.4.1
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
inflection==0.5.1
itemadapter==0.8.0
itemloaders==1.1.0
jmespath==1.0.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
lxml==5.2.1
multidict==6.0.5
mysqlclient==2.2.4
packaging==24.0
parsel==1.9.1
Protego==0.3.1
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyDispatcher==2.0.7
PyJWT==2.8.0
pyOpenSSL==24.1.0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
queuelib==1.6.2
referencing==0.35.1
requests==2.31.0
requests-file==2.0.0
rpds-py==0.18.1
Scrapy==2.11.1
scrapy-djangoitem==1.1.1
service-identity==24.1.0
shortuuid==1.0.13
six==1.16.0
sqlparse==0.5.0
tldextract==5.1.2
Twisted==24.3.0
# twisted-iocpsupport==1.0.4 # windows 전용
typing_extensions==4.11.0
tzdata==2024.1
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.2.1
w3lib==2.1.2
yarl==1.9.4
zope.interface==6.3
PyNaCl>=1.4.0,<2.0.0
gunicorn>=20.1.0,<21.0
Scrapy==2.11.1
